refactor(types): split types/index.ts into domain modules (#328)
- Create 11 domain modules: auth, common, crm, products, orders,
  wallet, admin, giftcard, email, vendor, contract
- Rewrite index.ts as pure re-export barrel (zero consumer changes)
- Fix duplicated ProductListParams (merged single declaration
  with full 9-field shape, GenericProductListParams extends it)
- Preserve all bilingual doc comments and exported symbols

TS: 0 errors | Tests: 71 files, 668 passed
